LANGCHAIN_MODELS/requirements.txt
# Langchain Core
Langchain
Langchain-Core
langchain-community
langchain-openai
duckduckgo-search
faiss-cpu
tiktoken

# OpenAI integration

Langchain-OpenAI
OpenAI

# Environment Variable Management

python-dotenv